requirements.txt
fastapi
uvicorn[standard]
requests
python-dateutil 
python-dotenv
pandas 
tensorflow
keras 